Boxplot Parameters of a matrix or data frame

This function calculates the statistics of the boxplot for the input matrix or data frame.

ds.boxplot(data, out.level=1.5, width = 0.15 , outl = TRUE, tojson=FALSE)

Arguments

data
The input numeric matrix or data frame.

out.level
Determines the length of the "whiskers" plot. If it is equal to zero or "outl" is set to F, no outliers will be returned.

width
The width level is determined 0.15 times the square root of the size of the input data.

outl
If TRUE the outliers will be computed at the selected "out.level" level (default is 1.5 times the Interquartile Range).

tojson
If TRUE the results are returned in json format

Value

Returns a list with the extracted components of ds.box for each variable/column of the input data.

Details

This function returns as a list object the statistical parameters needed to visualize boxplot.

See also

ds.box, ds.analysis, open_spending.ds

Examples

# with matrix as an input and the default parameters
Matrix <- cbind(Uni05 = (1:200)/21, Norm = rnorm(200),
         `5T` = rt(200, df = 5), Gam2 = rgamma(200, shape = 2))
ds.boxplot(Matrix, out.level=1.5, width = 0.15 , outl = TRUE, tojson=FALSE)

#> $Uni05
#> $Uni05$lo.whisker
#> [1] 0.04761905
#> 
#> $Uni05$lo.hinge
#> [1] 2.404762
#> 
#> $Uni05$median
#> [1] 4.785714
#> 
#> $Uni05$up.hinge
#> [1] 7.166667
#> 
#> $Uni05$up.whisker
#> [1] 9.52381
#> 
#> $Uni05$box.width
#> [1] 2.12
#> 
#> $Uni05$lo.out
#> numeric(0)
#> 
#> $Uni05$up.out
#> numeric(0)
#> 
#> $Uni05$n
#> [1] 200
#> 
#> 
#> $Norm
#> $Norm$lo.whisker
#> [1] -2.437264
#> 
#> $Norm$lo.hinge
#> [1] -0.5424603
#> 
#> $Norm$median
#> [1] 0.1151165
#> 
#> $Norm$up.hinge
#> [1] 0.7757827
#> 
#> $Norm$up.whisker
#> [1] 2.682557
#> 
#> $Norm$box.width
#> [1] 2.12
#> 
#> $Norm$lo.out
#> [1] -2.612334
#> 
#> $Norm$up.out
#> [1] 2.755418
#> 
#> $Norm$n
#> [1] 200
#> 
#> 
#> $`5T`
#> $`5T`$lo.whisker
#> [1] -2.713844
#> 
#> $`5T`$lo.hinge
#> [1] -0.7518927
#> 
#> $`5T`$median
#> [1] 0.005840514
#> 
#> $`5T`$up.hinge
#> [1] 0.6485875
#> 
#> $`5T`$up.whisker
#> [1] 2.740428
#> 
#> $`5T`$box.width
#> [1] 2.12
#> 
#> $`5T`$lo.out
#> [1] -4.634099
#> 
#> $`5T`$up.out
#> [1] 5.000904 3.585719
#> 
#> $`5T`$n
#> [1] 200
#> 
#> 
#> $Gam2
#> $Gam2$lo.whisker
#> [1] 0.05391239
#> 
#> $Gam2$lo.hinge
#> [1] 0.8693435
#> 
#> $Gam2$median
#> [1] 1.657201
#> 
#> $Gam2$up.hinge
#> [1] 2.458163
#> 
#> $Gam2$up.whisker
#> [1] 4.436686
#> 
#> $Gam2$box.width
#> [1] 2.12
#> 
#> $Gam2$lo.out
#> numeric(0)
#> 
#> $Gam2$up.out
#> [1] 5.537754 6.437423 5.609977 5.092422 5.898048 7.951376 4.906629 5.866432
#> 
#> $Gam2$n
#> [1] 200
#> 
#> 

# with data frame as an input, different parameters and json output
data <- iris
ds.boxplot(data, out.level=2, width = 0.25 , outl = FALSE, tojson=TRUE)

#> {"Sepal.Length":{"lo.whisker":[4.3],"lo.hinge":[5.1],"median":[5.8],"up.hinge":[6.4],"up.whisker":[7.9],"box.width":[3.06],"lo.out":{},"up.out":{},"n":[150]},"Sepal.Width":{"lo.whisker":[2],"lo.hinge":[2.8],"median":[3],"up.hinge":[3.3],"up.whisker":[4.2],"box.width":[3.06],"lo.out":{},"up.out":{},"n":[150]},"Petal.Length":{"lo.whisker":[1],"lo.hinge":[1.6],"median":[4.35],"up.hinge":[5.1],"up.whisker":[6.9],"box.width":[3.06],"lo.out":{},"up.out":{},"n":[150]},"Petal.Width":{"lo.whisker":[0.1],"lo.hinge":[0.3],"median":[1.3],"up.hinge":[1.8],"up.whisker":[2.5],"box.width":[3.06],"lo.out":{},"up.out":{},"n":[150]}} 

# OpenBudgets.eu Dataset Example:
Wuppertal_df

#>      ProduktNR   Kontotyp      Art Year       Amount ProduktbereichNR
#> 1      1109020     Ertrag Ergebnis 2009    203228.20               11
#> 2      1109020 Aufwendung Ergebnis 2009    219134.29               11
#> 3      3103040     Ertrag Ergebnis 2009   1926839.27               31
#> 4      3103040 Aufwendung Ergebnis 2009  11433218.77               31
#> 5      3105020     Ertrag Ergebnis 2009     18658.29               31
#> 6      3105020 Aufwendung Ergebnis 2009    147107.79               31
#> 7      1121010     Ertrag Ergebnis 2009    448699.77               11
#> 8      1121010 Aufwendung Ergebnis 2009    814812.26               11
#> 9      1211010     Ertrag Ergebnis 2009    296074.58               12
#> 10     1211010 Aufwendung Ergebnis 2009    315384.73               12
#> 11     2506020 Aufwendung Ergebnis 2009   2253840.64               25
#> 12     1116030     Ertrag Ergebnis 2009   6223634.83               11
#> 13     1116030 Aufwendung Ergebnis 2009   6698354.92               11
#> 14     1211050     Ertrag Ergebnis 2009       261.50               12
#> 15     1211050 Aufwendung Ergebnis 2009      2137.52               12
#> 16     1101010     Ertrag Ergebnis 2009   2739226.26               11
#> 17     1101010 Aufwendung Ergebnis 2009   2954238.51               11
#> 18     3603060     Ertrag Ergebnis 2009    717507.29               36
#> 19     3603060 Aufwendung Ergebnis 2009   1676459.33               36
#> 20     1105010     Ertrag Ergebnis 2009    578336.13               11
#> 21     1105010 Aufwendung Ergebnis 2009    624485.12               11
#> 22     1202010     Ertrag Ergebnis 2009     38363.53               12
#> 23     1202010 Aufwendung Ergebnis 2009    491108.84               12
#> 24     4101040     Ertrag Ergebnis 2009      1471.65               41
#> 25     4101040 Aufwendung Ergebnis 2009    711765.39               41
#> 26     5405010     Ertrag Ergebnis 2009  12588303.35               54
#> 27     5405010 Aufwendung Ergebnis 2009  13908794.00               54
#> 28     1212020     Ertrag Ergebnis 2009     68906.35               12
#> 29     1212020 Aufwendung Ergebnis 2009   1106200.81               12
#> 30     5503010     Ertrag Ergebnis 2009    453900.28               55
#> 31     5503010 Aufwendung Ergebnis 2009   2337694.31               55
#> 32     3605030     Ertrag Ergebnis 2009         8.40               36
#> 33     3605030 Aufwendung Ergebnis 2009   3651108.73               36
#> 34     1214010     Ertrag Ergebnis 2009  13434982.73               12
#> 35     1214010 Aufwendung Ergebnis 2009  13367695.14               12
#> 36     5602010     Ertrag Ergebnis 2009    774757.99               56
#> 37     5602010 Aufwendung Ergebnis 2009   2735288.77               56
#> 38     1113010     Ertrag Ergebnis 2009    526781.42               11
#> 39     1113010 Aufwendung Ergebnis 2009    567749.96               11
#> 40     5402020     Ertrag Ergebnis 2009    650188.26               54
#> 41     5402020 Aufwendung Ergebnis 2009   5740961.86               54
#> 42     4101050     Ertrag Ergebnis 2009    230404.16               41
#> 43     4101050 Aufwendung Ergebnis 2009   1312266.34               41
#> 44     1114020     Ertrag Ergebnis 2009    156902.50               11
#> 45     1114020 Aufwendung Ergebnis 2009    768988.54               11
#> 46     1110010     Ertrag Ergebnis 2009    359386.88               11
#> 47     1110010 Aufwendung Ergebnis 2009    385210.79               11
#> 48     2509010     Ertrag Ergebnis 2009     79431.97               25
#> 49     2509010 Aufwendung Ergebnis 2009   4979509.77               25
#> 50     1114010     Ertrag Ergebnis 2009    242046.07               11
#> 51     1114010 Aufwendung Ergebnis 2009    261390.58               11
#> 52     1207010     Ertrag Ergebnis 2009   1666708.07               12
#> 53     1207010 Aufwendung Ergebnis 2009   4271374.43               12
#> 54     3107020     Ertrag Ergebnis 2009    522605.82               31
#> 55     3107020 Aufwendung Ergebnis 2009    495379.14               31
#> 56     5701010     Ertrag Ergebnis 2009    251206.83               57
#> 57     5701010 Aufwendung Ergebnis 2009   1651086.60               57
#> 58     2104030     Ertrag Ergebnis 2009   4594934.08               21
#> 59     2104030 Aufwendung Ergebnis 2009     87150.01               21
#> 60     5402010     Ertrag Ergebnis 2009   2084484.47               54
#> 61     5402010     Ertrag Ergebnis 2009   2084484.47               54
#> 62     5402010 Aufwendung Ergebnis 2009   4903972.08               54
#> 63     5402010 Aufwendung Ergebnis 2009   4903972.08               54
#> 64     1117020     Ertrag Ergebnis 2009   4897334.96               11
#> 65     1117020 Aufwendung Ergebnis 2009   6273959.37               11
#> 66     5305010     Ertrag Ergebnis 2009    776329.01               53
#> 67     5305010 Aufwendung Ergebnis 2009    542479.50               53
#> 68     1103010     Ertrag Ergebnis 2009    215291.26               11
#> 69     1103010 Aufwendung Ergebnis 2009    231762.07               11
#> 70     5304010     Ertrag Ergebnis 2009    492147.77               53
#> 71     5304010 Aufwendung Ergebnis 2009    492122.68               53
#> 72     3603050     Ertrag Ergebnis 2009     39090.20               36
#> 73     3603050 Aufwendung Ergebnis 2009   1125002.61               36
#> 74     1128010     Ertrag Ergebnis 2009    118103.81               11
#> 75     1128010 Aufwendung Ergebnis 2009       340.92               11
#> 76     5301010     Ertrag Ergebnis 2009  20363833.63               53
#> 77     5301010 Aufwendung Ergebnis 2009   2820000.00               53
#> 78     1108010     Ertrag Ergebnis 2009    401485.18               11
#> 79     1108010 Aufwendung Ergebnis 2009    433601.81               11
#> 80     1203010     Ertrag Ergebnis 2009      1180.00               12
#> 81     1203010 Aufwendung Ergebnis 2009    206371.54               12
#> 82     5404010     Ertrag Ergebnis 2009    252818.84               54
#> 83     5404010 Aufwendung Ergebnis 2009   7459153.19               54
#> 84     4202010     Ertrag Ergebnis 2009    224099.36               42
#> 85     4202010 Aufwendung Ergebnis 2009   1627189.66               42
#> 86     1123010     Ertrag Ergebnis 2009   1168093.99               11
#> 87     1123010 Aufwendung Ergebnis 2009   1274402.01               11
#> 88     2104010     Ertrag Ergebnis 2009    737011.76               21
#> 89     2104010 Aufwendung Ergebnis 2009   2048393.26               21
#> 90     5703010     Ertrag Ergebnis 2009    123233.89               57
#> 91     5703010 Aufwendung Ergebnis 2009   4362977.89               57
#> 92     2504010     Ertrag Ergebnis 2009    659330.66               25
#> 93     2504010 Aufwendung Ergebnis 2009   6196642.52               25
#> 94     1112020     Ertrag Ergebnis 2009    338869.76               11
#> 95     1112020 Aufwendung Ergebnis 2009    366006.42               11
#> 96     1102010     Ertrag Ergebnis 2009    633825.73               11
#> 97     1102010 Aufwendung Ergebnis 2009    686577.75               11
#> 98     2502010     Ertrag Ergebnis 2009    400387.16               25
#> 99     2502010 Aufwendung Ergebnis 2009    520813.98               25
#> 100    3602010     Ertrag Ergebnis 2009    337914.93               36
#>      ProduktgruppeNR
#> 1               1109
#> 2               1109
#> 3               3103
#> 4               3103
#> 5               3105
#> 6               3105
#> 7               1121
#> 8               1121
#> 9               1211
#> 10              1211
#> 11              2506
#> 12              1116
#> 13              1116
#> 14              1211
#> 15              1211
#> 16              1101
#> 17              1101
#> 18              3603
#> 19              3603
#> 20              1105
#> 21              1105
#> 22              1202
#> 23              1202
#> 24              4101
#> 25              4101
#> 26              5405
#> 27              5405
#> 28              1212
#> 29              1212
#> 30              5503
#> 31              5503
#> 32              3605
#> 33              3605
#> 34              1214
#> 35              1214
#> 36              5602
#> 37              5602
#> 38              1113
#> 39              1113
#> 40              5402
#> 41              5402
#> 42              4101
#> 43              4101
#> 44              1114
#> 45              1114
#> 46              1110
#> 47              1110
#> 48              2509
#> 49              2509
#> 50              1114
#> 51              1114
#> 52              1207
#> 53              1207
#> 54              3107
#> 55              3107
#> 56              5701
#> 57              5701
#> 58              2104
#> 59              2104
#> 60              5402
#> 61              5402
#> 62              5402
#> 63              5402
#> 64              1117
#> 65              1117
#> 66              5305
#> 67              5305
#> 68              1103
#> 69              1103
#> 70              5304
#> 71              5304
#> 72              3603
#> 73              3603
#> 74              1128
#> 75              1128
#> 76              5301
#> 77              5301
#> 78              1108
#> 79              1108
#> 80              1203
#> 81              1203
#> 82              5404
#> 83              5404
#> 84              4202
#> 85              4202
#> 86              1123
#> 87              1123
#> 88              2104
#> 89              2104
#> 90              5703
#> 91              5703
#> 92              2504
#> 93              2504
#> 94              1112
#> 95              1112
#> 96              1102
#> 97              1102
#> 98              2502
#> 99              2502
#> 100             3602
#>                                                           Produkt
#> 1                     (entfallen in 2012) E-Government / Internet
#> 2                     (entfallen in 2012) E-Government / Internet
#> 3                                                 Nicht definiert
#> 4                                                 Nicht definiert
#> 5                        gesperrt: Hilfen nach dem SGB XII für Fl
#> 6                        gesperrt: Hilfen nach dem SGB XII für Fl
#> 7                                                      Postdienst
#> 8                                                      Postdienst
#> 9                                                  Bundestagswahl
#> 10                                                 Bundestagswahl
#> 11                                       Zuschuss für Tanztheater
#> 12                                         Beteiligungsmanagement
#> 13                                         Beteiligungsmanagement
#> 14                       Sonstige Wahlen u. sonstige Abstimmungen
#> 15                       Sonstige Wahlen u. sonstige Abstimmungen
#> 16                                   Rat und Gremien / Fraktionen
#> 17                                   Rat und Gremien / Fraktionen
#> 18                       Pl./Weiterentw. Freiräume u.Spielflächen
#> 19                       Pl./Weiterentw. Freiräume u.Spielflächen
#> 20                            GBL und Steuerungsunterstützung GB1
#> 21                            GBL und Steuerungsunterstützung GB1
#> 22                                   Öffentliche Toilettenanlagen
#> 23                                   Öffentliche Toilettenanlagen
#> 24                       Hilfe u.Beratung psychisch u.Suchtkranke
#> 25                       Hilfe u.Beratung psychisch u.Suchtkranke
#> 26                                               Straßenreinigung
#> 27                                               Straßenreinigung
#> 28                                             Gefahrenvorbeugung
#> 29                                             Gefahrenvorbeugung
#> 30                                      Wald- und Forstwirtschaft
#> 31                                      Wald- und Forstwirtschaft
#> 32                                                Nicht definiert
#> 33                                                Nicht definiert
#> 34                              Rettungsdienst (Gebührenhaushalt)
#> 35                              Rettungsdienst (Gebührenhaushalt)
#> 36                                                    Bodenschutz
#> 37                                                    Bodenschutz
#> 38                              Presse- und Öffentlichkeitsarbeit
#> 39                              Presse- und Öffentlichkeitsarbeit
#> 40                                                Nicht definiert
#> 41                                                Nicht definiert
#> 42                              Infektionsschutz u. QS Gesundheit
#> 43                              Infektionsschutz u. QS Gesundheit
#> 44                                          Kommunale Fortbildung
#> 45                                          Kommunale Fortbildung
#> 46                               Gleichstellung von Frau und Mann
#> 47                               Gleichstellung von Frau und Mann
#> 48                       Sammlung+son.Leist. Von der Heydt-Museum
#> 49                       Sammlung+son.Leist. Von der Heydt-Museum
#> 50                                                     Ausbildung
#> 51                                                     Ausbildung
#> 52                                       Einwohnerangelegenheiten
#> 53                                       Einwohnerangelegenheiten
#> 54                                                Nicht definiert
#> 55                                                Nicht definiert
#> 56                         Wirtschafts- und  Technologieförderung
#> 57                         Wirtschafts- und  Technologieförderung
#> 58                                                Nicht definiert
#> 59                                                Nicht definiert
#> 60                         Verkehrsleitende und -regelnde Anlagen
#> 61                                        Öffentliche Beleuchtung
#> 62                         Verkehrsleitende und -regelnde Anlagen
#> 63                                        Öffentliche Beleuchtung
#> 64                                                  Rechenzentrum
#> 65                                                  Rechenzentrum
#> 66                                        Kanalanschlussleitungen
#> 67                                        Kanalanschlussleitungen
#> 68                        Bürgerangelegenheiten und Bürgerservice
#> 69                        Bürgerangelegenheiten und Bürgerservice
#> 70                                 Koordinierung Abfallwirtschaft
#> 71                                 Koordinierung Abfallwirtschaft
#> 72                                                Nicht definiert
#> 73                                                Nicht definiert
#> 74                                            Personalnebenkosten
#> 75                                            Personalnebenkosten
#> 76                                     Energie- und Gasversorgung
#> 77                                     Energie- und Gasversorgung
#> 78                                                Nicht definiert
#> 79                                                Nicht definiert
#> 80                        Standesamtsaufsicht / Schiedsamtsangel.
#> 81                        Standesamtsaufsicht / Schiedsamtsangel.
#> 82                                Öffentlicher Personennahverkehr
#> 83                                Öffentlicher Personennahverkehr
#> 84                                                 Sportförderung
#> 85                                                 Sportförderung
#> 86                                                 Service-Center
#> 87                                                 Service-Center
#> 88                                                  Medienzentrum
#> 89                                                  Medienzentrum
#> 90                               Historische Stadthalle Wuppertal
#> 91                               Historische Stadthalle Wuppertal
#> 92                                       Medien und Informationen
#> 93                                       Medien und Informationen
#> 94                          Korruptionsbekämpfung und -prävention
#> 95                          Korruptionsbekämpfung und -prävention
#> 96                               Bezirksvertretungen und Ehrenamt
#> 97                               Bezirksvertretungen und Ehrenamt
#> 98                                              KULTURsekretariat
#> 99                                              KULTURsekretariat
#> 100                                                   Tagespflege
#>                         Produktbereich
#> 1                    Innere Verwaltung
#> 2                    Innere Verwaltung
#> 3                   Soziale Leistungen
#> 4                   Soziale Leistungen
#> 5                   Soziale Leistungen
#> 6                   Soziale Leistungen
#> 7                    Innere Verwaltung
#> 8                    Innere Verwaltung
#> 9               Sicherheit und Ordnung
#> 10              Sicherheit und Ordnung
#> 11             Kultur und Wissenschaft
#> 12                   Innere Verwaltung
#> 13                   Innere Verwaltung
#> 14              Sicherheit und Ordnung
#> 15              Sicherheit und Ordnung
#> 16                   Innere Verwaltung
#> 17                   Innere Verwaltung
#> 18   Kinder-, Jugend- u. Familienhilfe
#> 19   Kinder-, Jugend- u. Familienhilfe
#> 20                   Innere Verwaltung
#> 21                   Innere Verwaltung
#> 22              Sicherheit und Ordnung
#> 23              Sicherheit und Ordnung
#> 24                  Gesundheitsdienste
#> 25                  Gesundheitsdienste
#> 26       Verkehrsflächen/-anlagen,ÖPNV
#> 27       Verkehrsflächen/-anlagen,ÖPNV
#> 28              Sicherheit und Ordnung
#> 29              Sicherheit und Ordnung
#> 30        Natur- und Landschaftspflege
#> 31        Natur- und Landschaftspflege
#> 32   Kinder-, Jugend- u. Familienhilfe
#> 33   Kinder-, Jugend- u. Familienhilfe
#> 34              Sicherheit und Ordnung
#> 35              Sicherheit und Ordnung
#> 36                        Umweltschutz
#> 37                        Umweltschutz
#> 38                   Innere Verwaltung
#> 39                   Innere Verwaltung
#> 40       Verkehrsflächen/-anlagen,ÖPNV
#> 41       Verkehrsflächen/-anlagen,ÖPNV
#> 42                  Gesundheitsdienste
#> 43                  Gesundheitsdienste
#> 44                   Innere Verwaltung
#> 45                   Innere Verwaltung
#> 46                   Innere Verwaltung
#> 47                   Innere Verwaltung
#> 48             Kultur und Wissenschaft
#> 49             Kultur und Wissenschaft
#> 50                   Innere Verwaltung
#> 51                   Innere Verwaltung
#> 52              Sicherheit und Ordnung
#> 53              Sicherheit und Ordnung
#> 54                  Soziale Leistungen
#> 55                  Soziale Leistungen
#> 56            Wirtschaft und Tourismus
#> 57            Wirtschaft und Tourismus
#> 58                 Schulträgeraufgaben
#> 59                 Schulträgeraufgaben
#> 60       Verkehrsflächen/-anlagen,ÖPNV
#> 61       Verkehrsflächen/-anlagen,ÖPNV
#> 62       Verkehrsflächen/-anlagen,ÖPNV
#> 63       Verkehrsflächen/-anlagen,ÖPNV
#> 64                   Innere Verwaltung
#> 65                   Innere Verwaltung
#> 66                 Ver- und Entsorgung
#> 67                 Ver- und Entsorgung
#> 68                   Innere Verwaltung
#> 69                   Innere Verwaltung
#> 70                 Ver- und Entsorgung
#> 71                 Ver- und Entsorgung
#> 72   Kinder-, Jugend- u. Familienhilfe
#> 73   Kinder-, Jugend- u. Familienhilfe
#> 74                   Innere Verwaltung
#> 75                   Innere Verwaltung
#> 76                 Ver- und Entsorgung
#> 77                 Ver- und Entsorgung
#> 78                   Innere Verwaltung
#> 79                   Innere Verwaltung
#> 80              Sicherheit und Ordnung
#> 81              Sicherheit und Ordnung
#> 82       Verkehrsflächen/-anlagen,ÖPNV
#> 83       Verkehrsflächen/-anlagen,ÖPNV
#> 84                      Sportförderung
#> 85                      Sportförderung
#> 86                   Innere Verwaltung
#> 87                   Innere Verwaltung
#> 88                 Schulträgeraufgaben
#> 89                 Schulträgeraufgaben
#> 90            Wirtschaft und Tourismus
#> 91            Wirtschaft und Tourismus
#> 92             Kultur und Wissenschaft
#> 93             Kultur und Wissenschaft
#> 94                   Innere Verwaltung
#> 95                   Innere Verwaltung
#> 96                   Innere Verwaltung
#> 97                   Innere Verwaltung
#> 98             Kultur und Wissenschaft
#> 99             Kultur und Wissenschaft
#> 100  Kinder-, Jugend- u. Familienhilfe
#>                                                     Produktgruppe
#> 1                                   Geschäftsbereichsleitung GB 4
#> 2                                   Geschäftsbereichsleitung GB 4
#> 3                                           Grundsicherung SGB II
#> 4                                           Grundsicherung SGB II
#> 5                                             Hilfen nach AsylBLG
#> 6                                             Hilfen nach AsylBLG
#> 7                                                Zentrale Dienste
#> 8                                                Zentrale Dienste
#> 9                                                          Wahlen
#> 10                                                         Wahlen
#> 11                                           Theater und Konzerte
#> 12                            Finanzmanagement und Rechnungswesen
#> 13                            Finanzmanagement und Rechnungswesen
#> 14                                                         Wahlen
#> 15                                                         Wahlen
#> 16                                             Politische Gremien
#> 17                                             Politische Gremien
#> 18                                 Einrichtungen der Jugendarbeit
#> 19                                 Einrichtungen der Jugendarbeit
#> 20                                  Geschäftsbereichsleitung GB 1
#> 21                                  Geschäftsbereichsleitung GB 1
#> 22                                   Öffentliche Toilettenanlagen
#> 23                                   Öffentliche Toilettenanlagen
#> 24                        Aufgaben der unteren Gesundheitsbehörde
#> 25                        Aufgaben der unteren Gesundheitsbehörde
#> 26                              Straßenreinigung und Winterdienst
#> 27                              Straßenreinigung und Winterdienst
#> 28                                                    Brandschutz
#> 29                                                    Brandschutz
#> 30                                Wald, Forst- und Landwirtschaft
#> 31                                Wald, Forst- und Landwirtschaft
#> 32                       Leist. z.Förd.junger Menschen u.Familien
#> 33                       Leist. z.Förd.junger Menschen u.Familien
#> 34                                                 Rettungsdienst
#> 35                                                 Rettungsdienst
#> 36                                                    Bodenschutz
#> 37                                                    Bodenschutz
#> 38                              Presse- und Öffentlichkeitsarbeit
#> 39                              Presse- und Öffentlichkeitsarbeit
#> 40                                                Verkehrstechnik
#> 41                                                Verkehrstechnik
#> 42                        Aufgaben der unteren Gesundheitsbehörde
#> 43                        Aufgaben der unteren Gesundheitsbehörde
#> 44                           Personalfortbildung und -entwicklung
#> 45                           Personalfortbildung und -entwicklung
#> 46                               Gleichstellung von Frau und Mann
#> 47                               Gleichstellung von Frau und Mann
#> 48                                           Von der Heydt-Museum
#> 49                                           Von der Heydt-Museum
#> 50                           Personalfortbildung und -entwicklung
#> 51                           Personalfortbildung und -entwicklung
#> 52                              Einwohner- und Personenstandwesen
#> 53                              Einwohner- und Personenstandwesen
#> 54                                       Hilfen bei Wohnproblemen
#> 55                                       Hilfen bei Wohnproblemen
#> 56                                           Wirtschaftsförderung
#> 57                                           Wirtschaftsförderung
#> 58                                                  Medienzentrum
#> 59                                                  Medienzentrum
#> 60                                                Verkehrstechnik
#> 61                                                Verkehrstechnik
#> 62                                                Verkehrstechnik
#> 63                                                Verkehrstechnik
#> 64                              Technik. Informationsverarbeitung
#> 65                              Technik. Informationsverarbeitung
#> 66                                    Sonstige Abwasserwirtschaft
#> 67                                    Sonstige Abwasserwirtschaft
#> 68                                             Verwaltungsführung
#> 69                                             Verwaltungsführung
#> 70                                 Koordinierung Abfallwirtschaft
#> 71                                 Koordinierung Abfallwirtschaft
#> 72                                 Einrichtungen der Jugendarbeit
#> 73                                 Einrichtungen der Jugendarbeit
#> 74                                            Personalnebenkosten
#> 75                                            Personalnebenkosten
#> 76                                                     Versorgung
#> 77                                                     Versorgung
#> 78                                                           <NA>
#> 79                                                           <NA>
#> 80                        Standesamtsaufsicht / Schiedsamtsangel.
#> 81                        Standesamtsaufsicht / Schiedsamtsangel.
#> 82                                                           ÖPNV
#> 83                                                           ÖPNV
#> 84                                                 Sportförderung
#> 85                                                 Sportförderung
#> 86                                                 Service-Center
#> 87                                                 Service-Center
#> 88                                                  Medienzentrum
#> 89                                                  Medienzentrum
#> 90                                       Allgemeine Einrichtungen
#> 91                                       Allgemeine Einrichtungen
#> 92                                                     Bibliothek
#> 93                                                     Bibliothek
#> 94                                               Rechnungsprüfung
#> 95                                               Rechnungsprüfung
#> 96                                            Bezirksvertretungen
#> 97                                            Bezirksvertretungen
#> 98                                          KULTURsekretariat NRW
#> 99                                          KULTURsekretariat NRW
#> 100                                                   Tagespflege
#>  [ reached getOption("max.print") -- omitted 6125 rows ]

ds.boxplot(Wuppertal_df$Amount, out.level = 2.5, width = 0.15 , outl = TRUE, tojson = FALSE)

#> $data
#> $data$lo.whisker
#> [1] -2040681
#> 
#> $data$lo.hinge
#> [1] 243696.1
#> 
#> $data$median
#> [1] 736038.1
#> 
#> $data$up.hinge
#> [1] 2653000
#> 
#> $data$up.whisker
#> [1] 8674230
#> 
#> $data$box.width
#> [1] 11.83
#> 
#> $data$lo.out
#> numeric(0)
#> 
#> $data$up.out
#>   [1]  11433219  12588303  13908794  13434983  13367695  20363834  18596765
#>   [8]  38919541  12087507  52347353  33998451 123000513  21222047  10725693
#>  [15]  35751463  10341147  15254723  35090973  27082754  27146703  13157264
#>  [22]  13167243  37880258   9341974   9609349  19359142  33220817 109882054
#>  [29]  98946317  57662324  14111218  69430567  71373909  19809289  28196859
#>  [36]  13125331  16108616  12545655  13490920  20484759   9902481  14960053
#>  [43]  14120206  14235182  11618489  53727427  32266240 126131502  21566311
#>  [50]   9812885   9960402  28090219  28166391  18827216  11507974  14359599
#>  [57]  38461868   9396412  20507284  32293810 118183158 107031348   9335567
#>  [64]  62452663  48053894  69163751  11871817  16994157  16861927  40211120
#>  [71]   9785389  23363671  23703989   8957646  16774036  38397389 110490838
#>  [78]  98463131  68664650   9167526  27520836   9129925  12303206  12299368
#>  [85]   8811953  18710699  23690190  37382608  12083178  18657799   9134164
#>  [92]  46721277 125413158  11413752  50736849 313792985  87079489 377304133
#>  [99]  45220103  14698820  56394866   9779963   8935805  37885319  27807840
#> [106]  27344379  25199092  13654755  40005202  68664650  54068636  63789829
#> [113]  11444153  14893324   9520040   8936067  27766643  14219627  14270231
#> [120]  14559481  20169875  19465000   9298807  11528091  17153831  11624622
#> [127]  55127807  65966865 148072678  33398527   9509631  28758152  28678071
#> [134]  16852268  12834422  14851502  42919047   9362517  19954589  35681724
#> [141] 107325831  96344094  71683658  10647204  61619286  74017762  11520309
#> [148]  15147284  14298547  15423107  41462424  10791906   9151941  18486986
#> [155]  48871700 108987996  95606834  75976912   8886563  26979609   9254230
#> [162]  16704633  14933422  20839109  21573744  35682696  11528409  18516693
#> [169]   9384789  65518656 144587180  11697577 313114558  95098003 416056048
#> [176]  43022313  11601109  60111927   9447360   9707172  41214557  29059868
#> [183]  28027898  26118507  18888369  44649293  75976912  12094008  40925932
#> [190]  10694025  13821511  16911118   9573738  27312355  14219627  14270231
#> [197]  14559481  19692228  19465000   9281419  11511254  16997600  11842030
#> [204]  55551482  65250849 145945230  32570583   9594556  28758152  28682165
#> [211]  14944211  12867027  14734836  43069044   9467249  19698589  36069079
#> [218] 107325831  96350533  72870063  11596942  62753296  71221052  12133963
#> [225]  16707056  19516231  34945784  50742865   9556854  32738024 109494501
#> [232] 108982337   8951308  27803996   9831205   9518918  12182698  19913773
#> [239]  23525724  37908393  11248192  17316297  10298558  73994670 157281407
#> [246]  12389663 289168472  95819556 443820452  43490536  11718553  50406451
#> [253]   9212583   9070670  45789971  16350747  35936905  28755830  29325167
#> [260]  26472486  16328706  44591952  76894541  13459454  17988626  17068454
#> [267]   9601312  12988774  15291711  20019856   9061253  42488820 106600000
#> [274] 106600000  74205453  27098674   9527584   9645361  32516207  35473178
#> [281]  19620000  22793581  38257290  15700000  15700000  21942700  18767800
#> [288]  62096462 150366755  11976082 302630700  99531000 445668500  30729000
#> [295]  11539452  52112562   9606871  28577000  28505884  27210190  17145026
#> [302]  47359815  44358707  74205453  12515496  30567335  17883107  14753898
#> [309]  17208572  10795573   9524698  12082821  14624352  19552939  14274353
#> [316]  14967690  32037093 109330960 110619219   8712193  28185790  10291308
#> [323]  12602023   9124364  35437352  36771072  11812797  21330672  25827642
#> [330]  40194420  13000000  12086663  23921327  16787270  64852509 157258424
#> [337]  12604759 304915334  97770261 399611656  26659809  12082172  61017441
#> [344]   9002204  14795995  49379074  29309548  29309767  23814062  19620195
#> [351]  44988779  49403490  78822268  18197334  15321408  16024642   8908068
#> [358]   9638579  13103852  15445343  19248806   9059575  42488820 106600000
#> [365] 106600000  75322844  27335205   9599403   9645361  35223117  38180626
#> [372]  19620000  23299566  39270917  15700000  15700000   9203283  21498200
#> [379]  18582200  62096462 150316817  12078119 308053700 101300000 457670000
#> [386]  31591000  11515555  51961034   9693376  28577000  28508825  27363825
#> [393]  17570173  49410188  46383602  75322844  12463324  30191083  15424498
#> [400]  14753510  16983839  11018116   9618959  13103852  15445343  19998806
#> [407]   9059575  35488820 106600000 106600000  27335205   9599403   9645361
#> [414]  35223117  38180626  19620000  23299566  40270917  15700000  15700000
#> [421]   9203283  21498200  18582200  69656462 156316817  12078119 331003700
#> [428] 100400000 435655000  28755000  11515555  51961034   9693376  12463324
#> [435]  30191083  28577000  28508825  27363825  17570173  49410188  45483602
#> [442]  75322844  22924498  14753510  16983839  11018116   9618959  13135244
#> [449]  14468816  19522706   9907564   9076656  27073300  26132416  10602865
#> [456]  10200000  10200000  40111668  42680556  17000000  29463203  48579051
#> [463]  14000000  14000000   9668938  17602400  17313750  76783000 165526545
#> [470]  12588831 328411700 104340000  14298578 452672000  29800000  11854075
#> [477]  56146764   9758016   8847206  12851331  36907213  28720000  29901313
#> [484]  27892496  25618403  57395898  49222354  15719991  90615528  35663564
#> [491]  37130172  14955643  17408080   9883079   9829835  13192499  14551733
#> [498]  19727596  10038848   9065620  29073300  27895044  10662239  10200000
#> [505]  10200000  43616676  46180981  17000000  29903001  49230417  14000000
#> [512]  14000000   9841404  16703300  16900350  82438000 169526607  12656697
#> [519] 342011700 109500000  16907566 471217000  30700000  46929865   9670258
#> [526]   8806558  12751331  36174699  29150350  30329122  27999342  24110830
#> [533]  58117139  52304122  15722659  92026785  40368011  47907967  15105314
#> [540]  17358751  10029517   9857222  13289331  14634895  19784896  10178979
#> [547]   9059955  29073300  28492934  10719037  10200000  10200000  47519935
#> [554]  50104962  17000000  30375284  50248645  14000000  14000000   9964601
#> [561]  16447700  15454750  87228000 170526670  12720597 337100000 113800000
#> [568]   8700000   8700000  16972176 485776000  31450000  46551262   9713286
#> [575]   8866802  12653431  37574506  29590000  30761116  28044838  24007154
#> [582]  59669867  54378884  15726703  93593111  40367444  47910635  15152281
#> [589]  17270486  10179665   9887265  13410995  14733455  19912696  10293606
#> [596]   9045374  29073300  28613161  10766039  10200000  10200000  51692040
#> [603]  54317430  17000000  31128833  51574086  14000000  14000000  10109881
#> [610]  16012300  14770150  88582500 174096670  12797649 334600000 118300000
#> [617]  17047088 500075000  32250000  46580903   9777372   8966708  12653431
#> [624]  37691678  30030000  31204139  28135091  12000000  15219900  23496842
#> [631]  59788312  55378525  15536988  95008030  40367364  47922965  15150019
#> [638]  17209209  10266524   9964576  13534788  14839665  20041016  10408697
#> [645]   9031622  29073300  28772440  10833112  10200000  10200000  56244145
#> [652]  58912067  17000000  32256372  52911965  14000000  14000000  10381436
#> [659]  15643500  14743950  88582500 174176670  12878432 331000000 122900000
#> [666]  17125210 507995000  17750000  46642336   9850728   9070656  12653431
#> [673]  38961869  30480000  31652396  28277698  12000000  15268200  23336027
#> [680]  60466128  56409617  15536777  96640834  40367364  47936039  15148603
#> [687]  17269689  10355394   9948306
#> 
#> $data$n
#> [1] 6225
#> 
#>