Calculation of the Statistic Measures

This function calculates the basic descriptive measures of the input dataset.

ds.statistics(data, tojson=FALSE)

Arguments

data
A numeric vector, matrix or data frame

tojson
If TRUE the results are returned in json format, default returns a list

Value

A list or json file with the following components:

Min The minimum observed value of the input data

Max The maximum observed value of the input data

Range The range, defined as the difference of the maximum and the minimum value.

Mean The average value of the input data

Median The median value of the input data

Quantiles The 25% and 75% percentiles

Variance The variance of the input data

Standard Deviation The standard deviation of the input data

Skewness The Skewness of the input data

Kurtosis The Kurtosis of the input data

Details

This function returns the following values of the input data: minimum, maximum, range, mean, median, first and third quantiles, variance, standart deviation, skewness and kurtosis.

See also

open_spending.ds

Examples

# with matrix as an input and json outpout
Matrix <- cbind(Uni05 = (1:200)/21, Norm = rnorm(200),
        `5T` = rt(200, df = 5), Gam2 = rgamma(200, shape = 2))
ds.statistics(Matrix, tojson=TRUE)

#> {"Min":{"Uni05":[0.0476],"Norm":[-2.8218],"5T":[-4.4624],"Gam2":[0.0275]},"Max":{"Uni05":[9.5238],"Norm":[3.0081],"5T":[3.6418],"Gam2":[8.625]},"Range":{"Uni05":[9.4762],"Norm":[5.8299],"5T":[8.1042],"Gam2":[8.5975]},"Mean":{"Uni05":[4.7857],"Norm":[-0.0004],"5T":[-0.0112],"Gam2":[1.9207]},"Median":{"Uni05":[4.7857],"Norm":[0.0001],"5T":[0.0834],"Gam2":[1.5614]},"Quantiles":{"Uni05":[2.4167,7.1548],"Norm":[-0.687,0.7029],"5T":[-0.6672,0.7756],"Gam2":[0.912,2.69]},"Variance":{"Uni05":[7.5964],"Norm":[1.0494],"5T":[1.7732],"Gam2":[1.8683]},"StandardDeviation":{"Uni05":[2.7562],"Norm":[1.0244],"5T":[1.3316],"Gam2":[1.3668]},"Kurtosis":[-1.218,-0.1445,1.0644,2.5077],"Skewness":[1.3416e-016,-0.0745,-0.3999,1.302]} 

# with vector as an input
vec <- as.vector(iris$Sepal.Width)
ds.statistics(vec, tojson=FALSE)

#> $Min
#> $Min$data
#> [1] 2
#> 
#> 
#> $Max
#> $Max$data
#> [1] 4.4
#> 
#> 
#> $Range
#> $Range$data
#> [1] 2.4
#> 
#> 
#> $Mean
#> $Mean$data
#> [1] 3.057333
#> 
#> 
#> $Median
#> $Median$data
#> [1] 3
#> 
#> 
#> $Quantiles
#> $Quantiles$data
#> 25% 75% 
#> 2.8 3.3 
#> 
#> 
#> $Variance
#> $Variance$data
#> [1] 0.1899794
#> 
#> 
#> $StandardDeviation
#> $StandardDeviation$data
#> [1] 0.4358663
#> 
#> 
#> $Kurtosis
#>      data 
#> 0.1387047 
#> 
#> $Skewness
#>      data 
#> 0.3126147 
#> 

# with data frame as an input
ds.statistics(iris, tojson=FALSE)

#> $Min
#> $Min$Sepal.Length
#> [1] 4.3
#> 
#> $Min$Sepal.Width
#> [1] 2
#> 
#> $Min$Petal.Length
#> [1] 1
#> 
#> $Min$Petal.Width
#> [1] 0.1
#> 
#> 
#> $Max
#> $Max$Sepal.Length
#> [1] 7.9
#> 
#> $Max$Sepal.Width
#> [1] 4.4
#> 
#> $Max$Petal.Length
#> [1] 6.9
#> 
#> $Max$Petal.Width
#> [1] 2.5
#> 
#> 
#> $Range
#> $Range$Sepal.Length
#> [1] 3.6
#> 
#> $Range$Sepal.Width
#> [1] 2.4
#> 
#> $Range$Petal.Length
#> [1] 5.9
#> 
#> $Range$Petal.Width
#> [1] 2.4
#> 
#> 
#> $Mean
#> $Mean$Sepal.Length
#> [1] 5.843333
#> 
#> $Mean$Sepal.Width
#> [1] 3.057333
#> 
#> $Mean$Petal.Length
#> [1] 3.758
#> 
#> $Mean$Petal.Width
#> [1] 1.199333
#> 
#> 
#> $Median
#> $Median$Sepal.Length
#> [1] 5.8
#> 
#> $Median$Sepal.Width
#> [1] 3
#> 
#> $Median$Petal.Length
#> [1] 4.35
#> 
#> $Median$Petal.Width
#> [1] 1.3
#> 
#> 
#> $Quantiles
#> $Quantiles$Sepal.Length
#> 25% 75% 
#> 5.1 6.4 
#> 
#> $Quantiles$Sepal.Width
#> 25% 75% 
#> 2.8 3.3 
#> 
#> $Quantiles$Petal.Length
#> 25% 75% 
#> 1.6 5.1 
#> 
#> $Quantiles$Petal.Width
#> 25% 75% 
#> 0.3 1.8 
#> 
#> 
#> $Variance
#> $Variance$Sepal.Length
#> [1] 0.6856935
#> 
#> $Variance$Sepal.Width
#> [1] 0.1899794
#> 
#> $Variance$Petal.Length
#> [1] 3.116278
#> 
#> $Variance$Petal.Width
#> [1] 0.5810063
#> 
#> 
#> $StandardDeviation
#> $StandardDeviation$Sepal.Length
#> [1] 0.8280661
#> 
#> $StandardDeviation$Sepal.Width
#> [1] 0.4358663
#> 
#> $StandardDeviation$Petal.Length
#> [1] 1.765298
#> 
#> $StandardDeviation$Petal.Width
#> [1] 0.7622377
#> 
#> 
#> $Kurtosis
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>   -0.6058125    0.1387047   -1.4168574   -1.3581792 
#> 
#> $Skewness
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>    0.3086407    0.3126147   -0.2694109   -0.1009166 
#> 

# OpenBudgets.eu Dataset Example:
Wuppertal_df

#>      ProduktNR   Kontotyp      Art Year       Amount ProduktbereichNR
#> 1      1109020     Ertrag Ergebnis 2009    203228.20               11
#> 2      1109020 Aufwendung Ergebnis 2009    219134.29               11
#> 3      3103040     Ertrag Ergebnis 2009   1926839.27               31
#> 4      3103040 Aufwendung Ergebnis 2009  11433218.77               31
#> 5      3105020     Ertrag Ergebnis 2009     18658.29               31
#> 6      3105020 Aufwendung Ergebnis 2009    147107.79               31
#> 7      1121010     Ertrag Ergebnis 2009    448699.77               11
#> 8      1121010 Aufwendung Ergebnis 2009    814812.26               11
#> 9      1211010     Ertrag Ergebnis 2009    296074.58               12
#> 10     1211010 Aufwendung Ergebnis 2009    315384.73               12
#> 11     2506020 Aufwendung Ergebnis 2009   2253840.64               25
#> 12     1116030     Ertrag Ergebnis 2009   6223634.83               11
#> 13     1116030 Aufwendung Ergebnis 2009   6698354.92               11
#> 14     1211050     Ertrag Ergebnis 2009       261.50               12
#> 15     1211050 Aufwendung Ergebnis 2009      2137.52               12
#> 16     1101010     Ertrag Ergebnis 2009   2739226.26               11
#> 17     1101010 Aufwendung Ergebnis 2009   2954238.51               11
#> 18     3603060     Ertrag Ergebnis 2009    717507.29               36
#> 19     3603060 Aufwendung Ergebnis 2009   1676459.33               36
#> 20     1105010     Ertrag Ergebnis 2009    578336.13               11
#> 21     1105010 Aufwendung Ergebnis 2009    624485.12               11
#> 22     1202010     Ertrag Ergebnis 2009     38363.53               12
#> 23     1202010 Aufwendung Ergebnis 2009    491108.84               12
#> 24     4101040     Ertrag Ergebnis 2009      1471.65               41
#> 25     4101040 Aufwendung Ergebnis 2009    711765.39               41
#> 26     5405010     Ertrag Ergebnis 2009  12588303.35               54
#> 27     5405010 Aufwendung Ergebnis 2009  13908794.00               54
#> 28     1212020     Ertrag Ergebnis 2009     68906.35               12
#> 29     1212020 Aufwendung Ergebnis 2009   1106200.81               12
#> 30     5503010     Ertrag Ergebnis 2009    453900.28               55
#> 31     5503010 Aufwendung Ergebnis 2009   2337694.31               55
#> 32     3605030     Ertrag Ergebnis 2009         8.40               36
#> 33     3605030 Aufwendung Ergebnis 2009   3651108.73               36
#> 34     1214010     Ertrag Ergebnis 2009  13434982.73               12
#> 35     1214010 Aufwendung Ergebnis 2009  13367695.14               12
#> 36     5602010     Ertrag Ergebnis 2009    774757.99               56
#> 37     5602010 Aufwendung Ergebnis 2009   2735288.77               56
#> 38     1113010     Ertrag Ergebnis 2009    526781.42               11
#> 39     1113010 Aufwendung Ergebnis 2009    567749.96               11
#> 40     5402020     Ertrag Ergebnis 2009    650188.26               54
#> 41     5402020 Aufwendung Ergebnis 2009   5740961.86               54
#> 42     4101050     Ertrag Ergebnis 2009    230404.16               41
#> 43     4101050 Aufwendung Ergebnis 2009   1312266.34               41
#> 44     1114020     Ertrag Ergebnis 2009    156902.50               11
#> 45     1114020 Aufwendung Ergebnis 2009    768988.54               11
#> 46     1110010     Ertrag Ergebnis 2009    359386.88               11
#> 47     1110010 Aufwendung Ergebnis 2009    385210.79               11
#> 48     2509010     Ertrag Ergebnis 2009     79431.97               25
#> 49     2509010 Aufwendung Ergebnis 2009   4979509.77               25
#> 50     1114010     Ertrag Ergebnis 2009    242046.07               11
#> 51     1114010 Aufwendung Ergebnis 2009    261390.58               11
#> 52     1207010     Ertrag Ergebnis 2009   1666708.07               12
#> 53     1207010 Aufwendung Ergebnis 2009   4271374.43               12
#> 54     3107020     Ertrag Ergebnis 2009    522605.82               31
#> 55     3107020 Aufwendung Ergebnis 2009    495379.14               31
#> 56     5701010     Ertrag Ergebnis 2009    251206.83               57
#> 57     5701010 Aufwendung Ergebnis 2009   1651086.60               57
#> 58     2104030     Ertrag Ergebnis 2009   4594934.08               21
#> 59     2104030 Aufwendung Ergebnis 2009     87150.01               21
#> 60     5402010     Ertrag Ergebnis 2009   2084484.47               54
#> 61     5402010     Ertrag Ergebnis 2009   2084484.47               54
#> 62     5402010 Aufwendung Ergebnis 2009   4903972.08               54
#> 63     5402010 Aufwendung Ergebnis 2009   4903972.08               54
#> 64     1117020     Ertrag Ergebnis 2009   4897334.96               11
#> 65     1117020 Aufwendung Ergebnis 2009   6273959.37               11
#> 66     5305010     Ertrag Ergebnis 2009    776329.01               53
#> 67     5305010 Aufwendung Ergebnis 2009    542479.50               53
#> 68     1103010     Ertrag Ergebnis 2009    215291.26               11
#> 69     1103010 Aufwendung Ergebnis 2009    231762.07               11
#> 70     5304010     Ertrag Ergebnis 2009    492147.77               53
#> 71     5304010 Aufwendung Ergebnis 2009    492122.68               53
#> 72     3603050     Ertrag Ergebnis 2009     39090.20               36
#> 73     3603050 Aufwendung Ergebnis 2009   1125002.61               36
#> 74     1128010     Ertrag Ergebnis 2009    118103.81               11
#> 75     1128010 Aufwendung Ergebnis 2009       340.92               11
#> 76     5301010     Ertrag Ergebnis 2009  20363833.63               53
#> 77     5301010 Aufwendung Ergebnis 2009   2820000.00               53
#> 78     1108010     Ertrag Ergebnis 2009    401485.18               11
#> 79     1108010 Aufwendung Ergebnis 2009    433601.81               11
#> 80     1203010     Ertrag Ergebnis 2009      1180.00               12
#> 81     1203010 Aufwendung Ergebnis 2009    206371.54               12
#> 82     5404010     Ertrag Ergebnis 2009    252818.84               54
#> 83     5404010 Aufwendung Ergebnis 2009   7459153.19               54
#> 84     4202010     Ertrag Ergebnis 2009    224099.36               42
#> 85     4202010 Aufwendung Ergebnis 2009   1627189.66               42
#> 86     1123010     Ertrag Ergebnis 2009   1168093.99               11
#> 87     1123010 Aufwendung Ergebnis 2009   1274402.01               11
#> 88     2104010     Ertrag Ergebnis 2009    737011.76               21
#> 89     2104010 Aufwendung Ergebnis 2009   2048393.26               21
#> 90     5703010     Ertrag Ergebnis 2009    123233.89               57
#> 91     5703010 Aufwendung Ergebnis 2009   4362977.89               57
#> 92     2504010     Ertrag Ergebnis 2009    659330.66               25
#> 93     2504010 Aufwendung Ergebnis 2009   6196642.52               25
#> 94     1112020     Ertrag Ergebnis 2009    338869.76               11
#> 95     1112020 Aufwendung Ergebnis 2009    366006.42               11
#> 96     1102010     Ertrag Ergebnis 2009    633825.73               11
#> 97     1102010 Aufwendung Ergebnis 2009    686577.75               11
#> 98     2502010     Ertrag Ergebnis 2009    400387.16               25
#> 99     2502010 Aufwendung Ergebnis 2009    520813.98               25
#> 100    3602010     Ertrag Ergebnis 2009    337914.93               36
#>      ProduktgruppeNR
#> 1               1109
#> 2               1109
#> 3               3103
#> 4               3103
#> 5               3105
#> 6               3105
#> 7               1121
#> 8               1121
#> 9               1211
#> 10              1211
#> 11              2506
#> 12              1116
#> 13              1116
#> 14              1211
#> 15              1211
#> 16              1101
#> 17              1101
#> 18              3603
#> 19              3603
#> 20              1105
#> 21              1105
#> 22              1202
#> 23              1202
#> 24              4101
#> 25              4101
#> 26              5405
#> 27              5405
#> 28              1212
#> 29              1212
#> 30              5503
#> 31              5503
#> 32              3605
#> 33              3605
#> 34              1214
#> 35              1214
#> 36              5602
#> 37              5602
#> 38              1113
#> 39              1113
#> 40              5402
#> 41              5402
#> 42              4101
#> 43              4101
#> 44              1114
#> 45              1114
#> 46              1110
#> 47              1110
#> 48              2509
#> 49              2509
#> 50              1114
#> 51              1114
#> 52              1207
#> 53              1207
#> 54              3107
#> 55              3107
#> 56              5701
#> 57              5701
#> 58              2104
#> 59              2104
#> 60              5402
#> 61              5402
#> 62              5402
#> 63              5402
#> 64              1117
#> 65              1117
#> 66              5305
#> 67              5305
#> 68              1103
#> 69              1103
#> 70              5304
#> 71              5304
#> 72              3603
#> 73              3603
#> 74              1128
#> 75              1128
#> 76              5301
#> 77              5301
#> 78              1108
#> 79              1108
#> 80              1203
#> 81              1203
#> 82              5404
#> 83              5404
#> 84              4202
#> 85              4202
#> 86              1123
#> 87              1123
#> 88              2104
#> 89              2104
#> 90              5703
#> 91              5703
#> 92              2504
#> 93              2504
#> 94              1112
#> 95              1112
#> 96              1102
#> 97              1102
#> 98              2502
#> 99              2502
#> 100             3602
#>                                                           Produkt
#> 1                     (entfallen in 2012) E-Government / Internet
#> 2                     (entfallen in 2012) E-Government / Internet
#> 3                                                 Nicht definiert
#> 4                                                 Nicht definiert
#> 5                        gesperrt: Hilfen nach dem SGB XII für Fl
#> 6                        gesperrt: Hilfen nach dem SGB XII für Fl
#> 7                                                      Postdienst
#> 8                                                      Postdienst
#> 9                                                  Bundestagswahl
#> 10                                                 Bundestagswahl
#> 11                                       Zuschuss für Tanztheater
#> 12                                         Beteiligungsmanagement
#> 13                                         Beteiligungsmanagement
#> 14                       Sonstige Wahlen u. sonstige Abstimmungen
#> 15                       Sonstige Wahlen u. sonstige Abstimmungen
#> 16                                   Rat und Gremien / Fraktionen
#> 17                                   Rat und Gremien / Fraktionen
#> 18                       Pl./Weiterentw. Freiräume u.Spielflächen
#> 19                       Pl./Weiterentw. Freiräume u.Spielflächen
#> 20                            GBL und Steuerungsunterstützung GB1
#> 21                            GBL und Steuerungsunterstützung GB1
#> 22                                   Öffentliche Toilettenanlagen
#> 23                                   Öffentliche Toilettenanlagen
#> 24                       Hilfe u.Beratung psychisch u.Suchtkranke
#> 25                       Hilfe u.Beratung psychisch u.Suchtkranke
#> 26                                               Straßenreinigung
#> 27                                               Straßenreinigung
#> 28                                             Gefahrenvorbeugung
#> 29                                             Gefahrenvorbeugung
#> 30                                      Wald- und Forstwirtschaft
#> 31                                      Wald- und Forstwirtschaft
#> 32                                                Nicht definiert
#> 33                                                Nicht definiert
#> 34                              Rettungsdienst (Gebührenhaushalt)
#> 35                              Rettungsdienst (Gebührenhaushalt)
#> 36                                                    Bodenschutz
#> 37                                                    Bodenschutz
#> 38                              Presse- und Öffentlichkeitsarbeit
#> 39                              Presse- und Öffentlichkeitsarbeit
#> 40                                                Nicht definiert
#> 41                                                Nicht definiert
#> 42                              Infektionsschutz u. QS Gesundheit
#> 43                              Infektionsschutz u. QS Gesundheit
#> 44                                          Kommunale Fortbildung
#> 45                                          Kommunale Fortbildung
#> 46                               Gleichstellung von Frau und Mann
#> 47                               Gleichstellung von Frau und Mann
#> 48                       Sammlung+son.Leist. Von der Heydt-Museum
#> 49                       Sammlung+son.Leist. Von der Heydt-Museum
#> 50                                                     Ausbildung
#> 51                                                     Ausbildung
#> 52                                       Einwohnerangelegenheiten
#> 53                                       Einwohnerangelegenheiten
#> 54                                                Nicht definiert
#> 55                                                Nicht definiert
#> 56                         Wirtschafts- und  Technologieförderung
#> 57                         Wirtschafts- und  Technologieförderung
#> 58                                                Nicht definiert
#> 59                                                Nicht definiert
#> 60                         Verkehrsleitende und -regelnde Anlagen
#> 61                                        Öffentliche Beleuchtung
#> 62                         Verkehrsleitende und -regelnde Anlagen
#> 63                                        Öffentliche Beleuchtung
#> 64                                                  Rechenzentrum
#> 65                                                  Rechenzentrum
#> 66                                        Kanalanschlussleitungen
#> 67                                        Kanalanschlussleitungen
#> 68                        Bürgerangelegenheiten und Bürgerservice
#> 69                        Bürgerangelegenheiten und Bürgerservice
#> 70                                 Koordinierung Abfallwirtschaft
#> 71                                 Koordinierung Abfallwirtschaft
#> 72                                                Nicht definiert
#> 73                                                Nicht definiert
#> 74                                            Personalnebenkosten
#> 75                                            Personalnebenkosten
#> 76                                     Energie- und Gasversorgung
#> 77                                     Energie- und Gasversorgung
#> 78                                                Nicht definiert
#> 79                                                Nicht definiert
#> 80                        Standesamtsaufsicht / Schiedsamtsangel.
#> 81                        Standesamtsaufsicht / Schiedsamtsangel.
#> 82                                Öffentlicher Personennahverkehr
#> 83                                Öffentlicher Personennahverkehr
#> 84                                                 Sportförderung
#> 85                                                 Sportförderung
#> 86                                                 Service-Center
#> 87                                                 Service-Center
#> 88                                                  Medienzentrum
#> 89                                                  Medienzentrum
#> 90                               Historische Stadthalle Wuppertal
#> 91                               Historische Stadthalle Wuppertal
#> 92                                       Medien und Informationen
#> 93                                       Medien und Informationen
#> 94                          Korruptionsbekämpfung und -prävention
#> 95                          Korruptionsbekämpfung und -prävention
#> 96                               Bezirksvertretungen und Ehrenamt
#> 97                               Bezirksvertretungen und Ehrenamt
#> 98                                              KULTURsekretariat
#> 99                                              KULTURsekretariat
#> 100                                                   Tagespflege
#>                         Produktbereich
#> 1                    Innere Verwaltung
#> 2                    Innere Verwaltung
#> 3                   Soziale Leistungen
#> 4                   Soziale Leistungen
#> 5                   Soziale Leistungen
#> 6                   Soziale Leistungen
#> 7                    Innere Verwaltung
#> 8                    Innere Verwaltung
#> 9               Sicherheit und Ordnung
#> 10              Sicherheit und Ordnung
#> 11             Kultur und Wissenschaft
#> 12                   Innere Verwaltung
#> 13                   Innere Verwaltung
#> 14              Sicherheit und Ordnung
#> 15              Sicherheit und Ordnung
#> 16                   Innere Verwaltung
#> 17                   Innere Verwaltung
#> 18   Kinder-, Jugend- u. Familienhilfe
#> 19   Kinder-, Jugend- u. Familienhilfe
#> 20                   Innere Verwaltung
#> 21                   Innere Verwaltung
#> 22              Sicherheit und Ordnung
#> 23              Sicherheit und Ordnung
#> 24                  Gesundheitsdienste
#> 25                  Gesundheitsdienste
#> 26       Verkehrsflächen/-anlagen,ÖPNV
#> 27       Verkehrsflächen/-anlagen,ÖPNV
#> 28              Sicherheit und Ordnung
#> 29              Sicherheit und Ordnung
#> 30        Natur- und Landschaftspflege
#> 31        Natur- und Landschaftspflege
#> 32   Kinder-, Jugend- u. Familienhilfe
#> 33   Kinder-, Jugend- u. Familienhilfe
#> 34              Sicherheit und Ordnung
#> 35              Sicherheit und Ordnung
#> 36                        Umweltschutz
#> 37                        Umweltschutz
#> 38                   Innere Verwaltung
#> 39                   Innere Verwaltung
#> 40       Verkehrsflächen/-anlagen,ÖPNV
#> 41       Verkehrsflächen/-anlagen,ÖPNV
#> 42                  Gesundheitsdienste
#> 43                  Gesundheitsdienste
#> 44                   Innere Verwaltung
#> 45                   Innere Verwaltung
#> 46                   Innere Verwaltung
#> 47                   Innere Verwaltung
#> 48             Kultur und Wissenschaft
#> 49             Kultur und Wissenschaft
#> 50                   Innere Verwaltung
#> 51                   Innere Verwaltung
#> 52              Sicherheit und Ordnung
#> 53              Sicherheit und Ordnung
#> 54                  Soziale Leistungen
#> 55                  Soziale Leistungen
#> 56            Wirtschaft und Tourismus
#> 57            Wirtschaft und Tourismus
#> 58                 Schulträgeraufgaben
#> 59                 Schulträgeraufgaben
#> 60       Verkehrsflächen/-anlagen,ÖPNV
#> 61       Verkehrsflächen/-anlagen,ÖPNV
#> 62       Verkehrsflächen/-anlagen,ÖPNV
#> 63       Verkehrsflächen/-anlagen,ÖPNV
#> 64                   Innere Verwaltung
#> 65                   Innere Verwaltung
#> 66                 Ver- und Entsorgung
#> 67                 Ver- und Entsorgung
#> 68                   Innere Verwaltung
#> 69                   Innere Verwaltung
#> 70                 Ver- und Entsorgung
#> 71                 Ver- und Entsorgung
#> 72   Kinder-, Jugend- u. Familienhilfe
#> 73   Kinder-, Jugend- u. Familienhilfe
#> 74                   Innere Verwaltung
#> 75                   Innere Verwaltung
#> 76                 Ver- und Entsorgung
#> 77                 Ver- und Entsorgung
#> 78                   Innere Verwaltung
#> 79                   Innere Verwaltung
#> 80              Sicherheit und Ordnung
#> 81              Sicherheit und Ordnung
#> 82       Verkehrsflächen/-anlagen,ÖPNV
#> 83       Verkehrsflächen/-anlagen,ÖPNV
#> 84                      Sportförderung
#> 85                      Sportförderung
#> 86                   Innere Verwaltung
#> 87                   Innere Verwaltung
#> 88                 Schulträgeraufgaben
#> 89                 Schulträgeraufgaben
#> 90            Wirtschaft und Tourismus
#> 91            Wirtschaft und Tourismus
#> 92             Kultur und Wissenschaft
#> 93             Kultur und Wissenschaft
#> 94                   Innere Verwaltung
#> 95                   Innere Verwaltung
#> 96                   Innere Verwaltung
#> 97                   Innere Verwaltung
#> 98             Kultur und Wissenschaft
#> 99             Kultur und Wissenschaft
#> 100  Kinder-, Jugend- u. Familienhilfe
#>                                                     Produktgruppe
#> 1                                   Geschäftsbereichsleitung GB 4
#> 2                                   Geschäftsbereichsleitung GB 4
#> 3                                           Grundsicherung SGB II
#> 4                                           Grundsicherung SGB II
#> 5                                             Hilfen nach AsylBLG
#> 6                                             Hilfen nach AsylBLG
#> 7                                                Zentrale Dienste
#> 8                                                Zentrale Dienste
#> 9                                                          Wahlen
#> 10                                                         Wahlen
#> 11                                           Theater und Konzerte
#> 12                            Finanzmanagement und Rechnungswesen
#> 13                            Finanzmanagement und Rechnungswesen
#> 14                                                         Wahlen
#> 15                                                         Wahlen
#> 16                                             Politische Gremien
#> 17                                             Politische Gremien
#> 18                                 Einrichtungen der Jugendarbeit
#> 19                                 Einrichtungen der Jugendarbeit
#> 20                                  Geschäftsbereichsleitung GB 1
#> 21                                  Geschäftsbereichsleitung GB 1
#> 22                                   Öffentliche Toilettenanlagen
#> 23                                   Öffentliche Toilettenanlagen
#> 24                        Aufgaben der unteren Gesundheitsbehörde
#> 25                        Aufgaben der unteren Gesundheitsbehörde
#> 26                              Straßenreinigung und Winterdienst
#> 27                              Straßenreinigung und Winterdienst
#> 28                                                    Brandschutz
#> 29                                                    Brandschutz
#> 30                                Wald, Forst- und Landwirtschaft
#> 31                                Wald, Forst- und Landwirtschaft
#> 32                       Leist. z.Förd.junger Menschen u.Familien
#> 33                       Leist. z.Förd.junger Menschen u.Familien
#> 34                                                 Rettungsdienst
#> 35                                                 Rettungsdienst
#> 36                                                    Bodenschutz
#> 37                                                    Bodenschutz
#> 38                              Presse- und Öffentlichkeitsarbeit
#> 39                              Presse- und Öffentlichkeitsarbeit
#> 40                                                Verkehrstechnik
#> 41                                                Verkehrstechnik
#> 42                        Aufgaben der unteren Gesundheitsbehörde
#> 43                        Aufgaben der unteren Gesundheitsbehörde
#> 44                           Personalfortbildung und -entwicklung
#> 45                           Personalfortbildung und -entwicklung
#> 46                               Gleichstellung von Frau und Mann
#> 47                               Gleichstellung von Frau und Mann
#> 48                                           Von der Heydt-Museum
#> 49                                           Von der Heydt-Museum
#> 50                           Personalfortbildung und -entwicklung
#> 51                           Personalfortbildung und -entwicklung
#> 52                              Einwohner- und Personenstandwesen
#> 53                              Einwohner- und Personenstandwesen
#> 54                                       Hilfen bei Wohnproblemen
#> 55                                       Hilfen bei Wohnproblemen
#> 56                                           Wirtschaftsförderung
#> 57                                           Wirtschaftsförderung
#> 58                                                  Medienzentrum
#> 59                                                  Medienzentrum
#> 60                                                Verkehrstechnik
#> 61                                                Verkehrstechnik
#> 62                                                Verkehrstechnik
#> 63                                                Verkehrstechnik
#> 64                              Technik. Informationsverarbeitung
#> 65                              Technik. Informationsverarbeitung
#> 66                                    Sonstige Abwasserwirtschaft
#> 67                                    Sonstige Abwasserwirtschaft
#> 68                                             Verwaltungsführung
#> 69                                             Verwaltungsführung
#> 70                                 Koordinierung Abfallwirtschaft
#> 71                                 Koordinierung Abfallwirtschaft
#> 72                                 Einrichtungen der Jugendarbeit
#> 73                                 Einrichtungen der Jugendarbeit
#> 74                                            Personalnebenkosten
#> 75                                            Personalnebenkosten
#> 76                                                     Versorgung
#> 77                                                     Versorgung
#> 78                                                           <NA>
#> 79                                                           <NA>
#> 80                        Standesamtsaufsicht / Schiedsamtsangel.
#> 81                        Standesamtsaufsicht / Schiedsamtsangel.
#> 82                                                           ÖPNV
#> 83                                                           ÖPNV
#> 84                                                 Sportförderung
#> 85                                                 Sportförderung
#> 86                                                 Service-Center
#> 87                                                 Service-Center
#> 88                                                  Medienzentrum
#> 89                                                  Medienzentrum
#> 90                                       Allgemeine Einrichtungen
#> 91                                       Allgemeine Einrichtungen
#> 92                                                     Bibliothek
#> 93                                                     Bibliothek
#> 94                                               Rechnungsprüfung
#> 95                                               Rechnungsprüfung
#> 96                                            Bezirksvertretungen
#> 97                                            Bezirksvertretungen
#> 98                                          KULTURsekretariat NRW
#> 99                                          KULTURsekretariat NRW
#> 100                                                   Tagespflege
#>  [ reached getOption("max.print") -- omitted 6125 rows ]

ds.statistics(Wuppertal_df$Amount, tojson = TRUE)

#> {"Min":{"data":[-2040680.54]},"Max":{"data":[507995000]},"Range":{"data":[510035680.54]},"Mean":{"data":[6171229.3658]},"Median":{"data":[736038.09]},"Quantiles":{"data":[243696.13,2653000]},"Variance":{"data":[777106882358169]},"StandardDeviation":{"data":[27876636.8552]},"Kurtosis":[160.1519],"Skewness":[11.4762]}