Read and Calculate the Basic Information for Basic Descriptive Tasks from Open Spending and Rudolf APIs.

Extract and analyze the input data provided from Open Spending API of OpenBudgets.eu, using the ds.analysis function.

open_spending.ds(json_data, dimensions=NULL, amounts=NULL, measured.dimensions=NULL,
coef.outl=1.5, box.outliers=T, box.wdth=0.15,
cor.method= "pearson", freq.select=NULL)

Arguments

json_data
The json string, URL or file from Open Spending API

dimensions
The dimensions of the input data

amounts
The measures of the input data

measured.dimensions
The dimensions to which correspond amount/numeric variables

coef.outl
Determines the length of the "whiskers" plot. If it is equal to zero no outliers will be returned.

box.outliers
If TRUE the outliers will be computed at the selected "coef.outl" level (default is 1.5 times the Interquartile Range).

box.wdth
The width level is determined 0.15 times the square root of the size of the input data.

cor.method
The correlation coefficient method to compute: "pearson" (default), "kendall" or "spearman".

freq.select
One or more nominal variables to calculate their corresponding frequencies.

Value

A json string with the resulted parameters of the ds.analysis function.

Details

This function is used to read data in json format from Open Spending and Rudolf APIs., in order to implement some basic descriptive tasks through ds.analysis function.

See also

ds.analysis

Examples

# OpenBudgets.eu Dataset Example:
#data=Wuppertal_openspending
#open_spending.ds(data, dimensions ="functional_classification_3.Produktgruppe|date_2.Year",
  #              amounts = "Amount")